requirements.txt
google-genai
python-dotenv
openai
ddgs
pytz
openwakeword
faster-whisper
piper-tts
pyaudio
time